+++ temba/orgs/views.py
@@ -2052,8 +2052,8 @@ class OrgCRUDL(SmartCRUDL):
         def add_channel_section(self, formax, channel):
 
             if self.has_org_perm('channels.channel_read'):
-                from temba.channels.views import get_channel_icon, get_channel_read_url
-                formax.add_section('channel', get_channel_read_url(channel), icon=get_channel_icon(channel.channel_type), action='link')
+                from temba.channels.views import get_channel_read_url
+                formax.add_section('channel', get_channel_read_url(channel), icon=channel.get_type().icon, action='link')
 
         def derive_formax_sections(self, formax, context):